Successfully renames to com.moon_rocks_dev.aThousandWords, updates fileprovider path
