+++ app/src/androidTest/java/com/moon_rocks_dev/aThousandWords/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package io.rachelmunoz.imageThoughts;
+package com.moon_rocks_dev.aThousandWords;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
