+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="io.rachelmunoz.imageThoughts">
+          package="com.moon_rocks_dev.aThousandWords">
 
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="io.rachelmunoz.imageThoughts.ImageThoughtsListActivity">
+        <activity android:name="com.moon_rocks_dev.aThousandWords.ImageThoughtsListActivity">
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -20,11 +20,11 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name="io.rachelmunoz.imageThoughts.ImageThoughtsActivity">
+        <activity android:name="com.moon_rocks_dev.aThousandWords.ImageThoughtsActivity">
         </activity>
         <provider
             android:name="android.support.v4.content.FileProvider"
-            android:authorities="io.rachelmunoz.imageThoughts.fileprovider"
+            android:authorities="com.moon_rocks_dev.aThousandWords.fileprovider"
 
             android:exported="false"
             android:grantUriPermissions="true">
