+++ app/src/main/java/com/moon_rocks_dev/aThousandWords/DynamicRecyclerView.java
@@ -1,4 +1,4 @@
-package io.rachelmunoz.imageThoughts;
+package com.moon_rocks_dev.aThousandWords;
 
 import android.support.annotation.LayoutRes;
