+++ app/src/main/java/com/moon_rocks_dev/aThousandWords/ImageThoughtsListActivity.java
@@ -1,4 +1,4 @@
-package io.rachelmunoz.imageThoughts;
+package com.moon_rocks_dev.aThousandWords;
 
 import android.content.Intent;
 import android.support.v4.app.Fragment;
