+++ app/src/main/java/com/moon_rocks_dev/aThousandWords/database/ImageThoughtBaseHelper.java
@@ -1,10 +1,10 @@
-package io.rachelmunoz.imageThoughts.database;
+package com.moon_rocks_dev.aThousandWords.database;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import io.rachelmunoz.imageThoughts.database.ImageThoughtDbSchema.ImageThoughtTable;
+import com.moon_rocks_dev.aThousandWords.database.ImageThoughtDbSchema.ImageThoughtTable;
 
 /**
  * Created by rachelmunoz on 7/23/17.
