+++ app/src/main/java/com/moon_rocks_dev/aThousandWords/database/ImageThoughtDbSchema.java
@@ -1,4 +1,4 @@
-package io.rachelmunoz.imageThoughts.database;
+package com.moon_rocks_dev.aThousandWords.database;
 
 /**
  * Created by rachelmunoz on 7/23/17.
